Now, doATurn calls player's eachTurn explicitly, and player's eachTurn calls doATurn explicitly.
added a placeRect function to levelpainter; should probably add a setinterval after which we clear the tile, or something.

Added utils.getTilOccupants, which (in theory) gets the tile that's at x,y and returns a stringof its occupants.

Added some player is alive logic (that doesn't do anything yet) and a mousemove listener using jquery to (eventually) be ableto update the 'what's in this tile?' div even though we're not using the DOM anymore.